Move Email to domain/email package
